Develop ReactorManager module

+++ src/reactor_manager.cc
@@ -60,7 +60,7 @@ ServerReactorManager::ServerReactorManager(const NetAddress& listen_address,
     LOG(logger::kError, "Fail to init the acceptor!!!");
     ::exit(-1);
   }
-  for (int i = 1; i < io_thread_amount; ++i) {
+  for (size_t i = 1; i < io_thread_amount; ++i) {
     event_managers_.emplace_back(new EventManager);
   }
   balancer_ = std::make_unique<Balancer>(&event_managers_, 0);
